api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dill==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
flake8==7.2.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pytest==8.4.0
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-jose==3.5.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tomlkit==0.13.3
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
